lib: pull ecosystem post matching into util
