+++ lib/lib.js
@@ -3,7 +3,7 @@ import { join, parse } from "path";
 import matter from "gray-matter";
 import toml from "@iarna/toml";
 import { DateTime } from "luxon";
-import { getPostBySlug, getPostSlugs } from "@urbit/foundation-design-system";
+import { Markdown, getPostBySlug, getPostSlugs } from "@urbit/foundation-design-system";
 
 // NB Gavin: This could be much simpler if we converted everything to a more common standard, like YAML
 const options = {
@@ -166,3 +166,86 @@ export const pair = (arr) =>
       rows
     );
   }, []);
+
+// Match ecosystem report to posts to inject metadata
+export const matchEcosystemPost = (ecosystem) => {
+  const articles = getAllPosts(
+    ["title", "image", "date", "podcast", "slug"],
+    "articles",
+    "date"
+  );
+
+  const marketplaces = getAllPosts(["title", "image", "slug"], "marketplaces");
+  const podcasts = getAllPosts(
+    ["title", "image", "date", "podcast", "slug"],
+    "podcasts",
+    "date"
+  );
+  const organizations = getAllPosts(
+    ["title", "image", "slug"],
+    "organizations"
+  );
+  const applications = fs
+    .readdirSync(join(process.cwd(), "content/applications"), {
+      withFileTypes: true,
+    })
+    .filter((f) => f.isDirectory())
+    .map((dir) =>
+      getAllPosts(
+        ["title", "bgColor", "image", "slug", "description"],
+        `applications/${dir.name}`
+      )
+        .map((e) => ({ ...e, ship: dir.name }))
+        .flat()
+    )
+    .flat()
+    .sort((a, b) => {
+      const nameA = a.title.toLowerCase();
+      const nameB = b.title.toLowerCase();
+      return nameA < nameB ? -1 : 1;
+    });
+
+  const groups = fs
+    .readdirSync(join(process.cwd(), "content/groups"), {
+      withFileTypes: true,
+    })
+    .filter((f) => f.isDirectory())
+    .map((dir) =>
+      getAllPosts(
+        ["title", "tile", "slug", "description"],
+        `groups/${dir.name}`
+      )
+        .map((e) => ({ ...e, ship: dir.name }))
+        .flat()
+    )
+    .flat()
+    .sort((a, b) => {
+      const nameA = a.title.toLowerCase();
+      const nameB = b.title.toLowerCase();
+      return nameA < nameB ? -1 : 1;
+    });
+
+  ["featured-1", "featured-2", "featured-3"].forEach((feat) => {
+    if (ecosystem?.[feat]) {
+      const matchedPost = [
+        ...articles.map((e) => ({ ...e, type: "Article" })),
+        ...applications.map((e) => ({ ...e, type: "Application" })),
+        ...organizations.map((e) => ({ ...e, type: "Organization" })),
+        ...podcasts.map((e) => ({ ...e, type: "Podcast" })),
+        ...marketplaces.map((e) => ({ ...e, type: "Marketplace" })),
+        ...groups.map((e) => ({ ...e, type: "Group" }))
+      ].filter((e) => e.title === ecosystem[feat].title)?.[0];
+      ecosystem[feat].image =
+        ecosystem[feat].image || matchedPost?.image || null;
+      ecosystem[feat].type = matchedPost?.type || "Podcast";
+      ecosystem[feat].matchedPost = matchedPost || null;
+      ecosystem[feat].content = JSON.stringify(
+        Markdown.parse({
+          post: { content: ecosystem[feat].content },
+        })
+      );
+    }
+  });
+
+  return ecosystem;
+}
