+++ pages/ecosystem/spotlight/[slug].js
@@ -6,7 +6,6 @@ import {
   Container,
   SingleColumn,
   Section,
-  Markdown,
   IntraNav,
   getAllPosts,
   getPostBySlug,
@@ -14,8 +13,7 @@ import {
 import Header from "../../../components/Header";
 import Footer from "../../../components/Footer";
 import Post from "../../../components/ecosystem/Post";
-import fs from "fs";
-import path from "path";
+import { matchEcosystemPost } from "../../../lib/lib";
 
 export default function BasicPage({ post, markdown, search, index = false }) {
   const router = useRouter();
@@ -44,67 +42,11 @@ export default function BasicPage({ post, markdown, search, index = false }) {
 }
 
 export async function getStaticProps({ params }) {
-  const post = getPostBySlug(
+  const post = matchEcosystemPost(getPostBySlug(
     params.slug,
     ["title", "slug", "date", "featured-1", "featured-2", "featured-3"],
     "ecosystem/spotlight"
-  );
-  const marketplaces = getAllPosts(["title", "image", "slug"], "marketplaces");
-  const groups = getAllPosts(["title", "tile,", "description", "slug"], "groups");
-  const podcasts = getAllPosts(
-    ["title", "image", "date", "podcast", "slug"],
-    "podcasts",
-    "date"
-  );
-  const articles = getAllPosts(
-    ["title", "image", "date", "publication", "author", "slug"],
-    "articles",
-    "date"
-  );
-  const organizations = getAllPosts(
-    ["title", "image", "slug"],
-    "organizations"
-  );
-  const applications = fs
-    .readdirSync(path.join(process.cwd(), "content/applications"), {
-      withFileTypes: true,
-    })
-    .filter((f) => f.isDirectory())
-    .map((dir) =>
-      getAllPosts(
-        ["title", "bgColor", "image", "slug", "description"],
-        `applications/${dir.name}`
-      )
-        .map((e) => ({ ...e, ship: dir.name }))
-        .flat()
-    )
-    .flat()
-    .sort((a, b) => {
-      const nameA = a.title.toLowerCase();
-      const nameB = b.title.toLowerCase();
-      return nameA < nameB ? -1 : 1;
-    });
-
-  ["featured-1", "featured-2", "featured-3"].forEach((feat) => {
-    if (post?.[feat]) {
-      const matchedPost = [
-        ...applications.map((e) => ({ ...e, type: "Application" })),
-        ...organizations.map((e) => ({ ...e, type: "Organization" })),
-        ...groups.map((e) => ({ ...e, type: "Group" })),
-        ...podcasts.map((e) => ({ ...e, type: "Podcast" })),
-        ...marketplaces.map((e) => ({ ...e, type: "Marketplace" })),
-        ...articles.map((e) => ({ ...e, type: "Article" })),
-      ].filter((e) => e.title === post[feat].title)?.[0];
-      post[feat].image = post[feat]?.image || matchedPost?.image || null;
-      post[feat].type = matchedPost?.type || "Podcast";
-      post[feat].matchedPost = matchedPost || null;
-      post[feat].content = JSON.stringify(
-        Markdown.parse({
-          post: { content: post[feat].content },
-        })
-      );
-    }
-  });
+  ));
 
   return {
     props: { post },
